PVP-230 Introduce ButtonConfirm for consistent confirmation dialog
